refactor(auth): move cookie helpers private
